--- module_1/gradle/by.epam.module1.gradle.multi-project/api/src/main/java/App.java
@@ -1,5 +1,0 @@
-public class App {
-    public static void main(String[] args) {
-        System.out.println(Utils.isAllPositiveNumbers("12", "79"));
-    }
-}
