--- module_1/gradle/by.epam.module1.gradle.multi-project/build.gradle
@@ -1,15 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-group 'by.epam.module1.gradle'
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-}
