--- module_1/gradle/by.epam.module1.gradle.multi-project/core/build.gradle
@@ -1,18 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-group 'by.epam.module1.gradle'
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenLocal()
-    mavenCentral()
-}
-
-dependencies {
-    implementation group: 'by.epam.module1.gradle' , name: 'utils', version: '1.3.5'
-
-}
