--- module_1/gradle/by.epam.module1.gradle.multi-project/core/src/main/java/Utils.java
@@ -1,10 +1,0 @@
-public class Utils {
-    public static boolean isAllPositiveNumbers(String... str) {
-        for (String st : str) {
-            if (!StringUtils.isPositiveNumber(st)) {
-                return false;
-            }
-        }
-        return true;
-    }
-}
