--- module_1/gradle/by.epam.module1.gradle.multi-project/settings.gradle
@@ -1,4 +1,0 @@
-rootProject.name = 'by.epam.module1.gradle.multi-project'
-include 'api'
-include 'core'
-
