--- module_1/gradle/by.epam.module1.gradle.utils/.gitignore
@@ -1,3 +1,0 @@
-.idea
-.gradle
-build
