--- module_1/gradle/by.epam.module1.gradle.utils/build.gradle
@@ -1,33 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'maven-publish'
-}
-
-
-group 'by.epam.module1.gradle'
-version '1.3.5'
-archivesBaseName = 'utils'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenLocal()
-    mavenCentral()
-}
-
-publishing {
-    publications {
-        mavenJava(MavenPublication) {
-            from components.java
-        }
-    }
-}
-
-
-dependencies {
-
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
-    compile 'org.apache.commons:commons-lang3:3.10'
-
-}
-
