--- module_1/gradle/by.epam.module1.gradle.utils/src/main/java/StringUtils.java
@@ -1,7 +1,0 @@
-import static org.apache.commons.lang3.StringUtils.isNumeric;
-
-public class StringUtils {
-    public static boolean isPositiveNumber (String str) {
-        return isNumeric(str);
-    }
-}
