+++ module_1/gradle/multi-project/.gitignore
@@ -1,0 +1,3 @@
+.idea
+.gradle
+build
