+++ module_1/gradle/multi-project/api/src/main/java/com/epam/mjc/App.java
@@ -1,0 +1,7 @@
+package com.epam.mjc;
+
+public class App {
+    public static void main(String[] args) {
+        System.out.println(Utils.isAllPositiveNumbers("12", "79"));
+    }
+}
