+++ module_1/gradle/multi-project/core/src/main/java/com/epam/mjc/Utils.java
@@ -1,0 +1,12 @@
+package com.epam.mjc;
+
+public class Utils {
+    public static boolean isAllPositiveNumbers(String... str) {
+        for (String it : str) {
+            if (!StringUtils.isPositiveNumber(it)) {
+                return false;
+            }
+        }
+        return true;
+    }
+}
