+++ module_1/gradle/multi-project/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'by.epam.module1.gradle.multi-project'
+include 'api'
+include 'core'
