+++ module_1/gradle/utils/.gitignore
@@ -1,0 +1,3 @@
+.idea
+.gradle
+build
