+++ module_1/gradle/utils/src/main/java/com/epam/mjc/StringUtils.java
@@ -1,0 +1,12 @@
+package com.epam.mjc;
+
+import org.apache.commons.lang3.math.NumberUtils;
+
+public class StringUtils {
+    public static boolean isPositiveNumber(String str) {
+        if (NumberUtils.isCreatable(str)) {
+            return Double.parseDouble(str) > 0;
+        }
+        return false;
+    }
+}
