+++ module_1/gradle/utils/src/test/java/com/epam/mjc/StringUtilsTest.java
@@ -1,0 +1,25 @@
+package com.epam.mjc;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+class StringUtilsTest {
+
+    @Test
+    void isTestPositiveNumber() {
+        boolean result = StringUtils.isPositiveNumber("1.2");
+        Assertions.assertTrue(result);
+    }
+
+    @Test
+    void isTestNegativeNumber() {
+        boolean result = StringUtils.isPositiveNumber("-12");
+        Assertions.assertFalse(result);
+    }
+
+    @Test
+    void isTestSomeWord() {
+        boolean result = StringUtils.isPositiveNumber("sd");
+        Assertions.assertFalse(result);
+    }
+}
